@@ -2,7 +2,7 @@ use actix_web::web::{Data, Json, Query};
 use lemmy_api_common::{
   comment::{ListCommentReports, ListCommentReportsResponse},
   context::LemmyContext,
-  utils::check_community_mod_action_opt,
+  utils::check_community_mod_of_any_or_admin_action,
 };
 use lemmy_db_views::{comment_report_view::CommentReportQuery, structs::LocalUserView};
 use lemmy_utils::error::LemmyError;
@@ -18,7 +18,7 @@ pub async fn list_comment_reports(
   let community_id = data.community_id;
   let unresolved_only = data.unresolved_only.unwrap_or_default();
 
-  check_community_mod_action_opt(&local_user_view, community_id, &mut context.pool()).await?;
+  check_community_mod_of_any_or_admin_action(&local_user_view, &mut context.pool()).await?;
 
   let page = data.page;
   let limit = data.limit;
