@@ -2,7 +2,7 @@ use actix_web::web::{Data, Json, Query};
 use lemmy_api_common::{
   context::LemmyContext,
   person::{GetReportCount, GetReportCountResponse},
-  utils::check_community_mod_action_opt,
+  utils::check_community_mod_of_any_or_admin_action,
 };
 use lemmy_db_views::structs::{
   CommentReportView,
@@ -22,7 +22,7 @@ pub async fn report_count(
   let admin = local_user_view.local_user.admin;
   let community_id = data.community_id;
 
-  check_community_mod_action_opt(&local_user_view, community_id, &mut context.pool()).await?;
+  check_community_mod_of_any_or_admin_action(&local_user_view, &mut context.pool()).await?;
 
   let comment_reports =
     CommentReportView::get_report_count(&mut context.pool(), person_id, admin, community_id)
