@@ -4,9 +4,10 @@ use crate::schema::local_user_language;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[cfg_attr(feature = "full", derive(Queryable, Identifiable))]
+#[cfg_attr(feature = "full", derive(Queryable, Selectable, Identifiable))]
 #[cfg_attr(feature = "full", diesel(table_name = local_user_language))]
 #[cfg_attr(feature = "full", diesel(primary_key(local_user_id, language_id)))]
+#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
 pub struct LocalUserLanguage {
   pub local_user_id: LocalUserId,
   pub language_id: LanguageId,
@@ -24,9 +25,10 @@ pub struct LocalUserLanguageForm {
 use crate::schema::community_language;
 
 #[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[cfg_attr(feature = "full", derive(Queryable, Identifiable))]
+#[cfg_attr(feature = "full", derive(Queryable, Selectable, Identifiable))]
 #[cfg_attr(feature = "full", diesel(table_name = community_language))]
 #[cfg_attr(feature = "full", diesel(primary_key(community_id, language_id)))]
+#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
 pub struct CommunityLanguage {
   pub community_id: CommunityId,
   pub language_id: LanguageId,
@@ -44,9 +46,10 @@ pub struct CommunityLanguageForm {
 use crate::schema::site_language;
 
 #[derive(Clone, PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[cfg_attr(feature = "full", derive(Queryable, Identifiable))]
+#[cfg_attr(feature = "full", derive(Queryable, Selectable, Identifiable))]
 #[cfg_attr(feature = "full", diesel(table_name = site_language))]
 #[cfg_attr(feature = "full", diesel(primary_key(site_id, language_id)))]
+#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
 pub struct SiteLanguage {
   pub site_id: SiteId,
   pub language_id: LanguageId,
