@@ -4,13 +4,14 @@ use crate::schema::email_verification;
 use chrono::{DateTime, Utc};
 
 #[derive(Clone)]
-#[cfg_attr(feature = "full", derive(Queryable, Identifiable))]
+#[cfg_attr(feature = "full", derive(Queryable, Selectable, Identifiable))]
 #[cfg_attr(feature = "full", diesel(table_name = email_verification))]
+#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
 pub struct EmailVerification {
   pub id: i32,
   pub local_user_id: LocalUserId,
   pub email: String,
-  pub verification_code: String,
+  pub verification_token: String,
   pub published: DateTime<Utc>,
 }
 
