@@ -11,7 +11,10 @@ use typed_builder::TypedBuilder;
 
 #[skip_serializing_none]
 #[derive(PartialEq, Eq, Debug, Clone, Serialize, Deserialize)]
-#[cfg_attr(feature = "full", derive(Queryable, Associations, Identifiable, TS))]
+#[cfg_attr(
+  feature = "full",
+  derive(Queryable, Selectable, Associations, Identifiable, TS)
+)]
 #[cfg_attr(feature = "full", diesel(table_name = image_upload))]
 #[cfg_attr(feature = "full", diesel(primary_key(pictrs_alias)))]
 #[cfg_attr(feature = "full", ts(export))]
@@ -19,6 +22,7 @@ use typed_builder::TypedBuilder;
   feature = "full",
   diesel(belongs_to(crate::source::local_user::LocalUser))
 )]
+#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
 pub struct ImageUpload {
   pub local_user_id: LocalUserId,
   pub pictrs_alias: String,
