@@ -9,12 +9,16 @@ use ts_rs::TS;
 
 #[skip_serializing_none]
 #[derive(PartialEq, Eq, Serialize, Deserialize, Debug, Clone)]
-#[cfg_attr(feature = "full", derive(Queryable, Associations, Identifiable, TS))]
+#[cfg_attr(
+  feature = "full",
+  derive(Queryable, Selectable, Associations, Identifiable, TS)
+)]
 #[cfg_attr(
   feature = "full",
   diesel(belongs_to(crate::source::private_message::PrivateMessage))
 )]
 #[cfg_attr(feature = "full", diesel(table_name = private_message_report))]
+#[cfg_attr(feature = "full", diesel(check_for_backend(diesel::pg::Pg)))]
 #[cfg_attr(feature = "full", ts(export))]
 /// The private message report.
 pub struct PrivateMessageReport {
