@@ -49,8 +49,13 @@ fn queries<'a>() -> Queries<
   let list = move |mut conn: DbConn<'a>, options: RegistrationApplicationQuery| async move {
     let mut query = all_joins(registration_application::table.into_boxed());
 
+    // If viewing all applications, order by newest, but if viewing unresolved only, show the oldest first (FIFO)
     if options.unread_only {
-      query = query.filter(registration_application::admin_id.is_null())
+      query = query
+        .filter(registration_application::admin_id.is_null())
+        .order_by(registration_application::published.asc());
+    } else {
+      query = query.order_by(registration_application::published.desc());
     }
 
     if options.verified_email_only {
@@ -59,10 +64,7 @@ fn queries<'a>() -> Queries<
 
     let (limit, offset) = limit_and_offset(options.page, options.limit)?;
 
-    query = query
-      .limit(limit)
-      .offset(offset)
-      .order_by(registration_application::published.asc());
+    query = query.limit(limit).offset(offset);
 
     query.load::<RegistrationApplicationView>(&mut conn).await
   };
